Mostrando mensagens

+++ orm/crud.php
@@ -1,7 +1,7 @@
 <?php
 
-require_once 'conexao.php';
-require_once 'src/Product.php';
+require_once 'bootstrap.php';
+require_once 'src/simples/Product.php';
 
 // http://doctrine-orm.readthedocs.org/en/latest/tutorials/getting-started.html#starting-with-the-product
 
@@ -12,17 +12,44 @@ $entityManager->persist($produto);
 $entityManager->flush();
 
 $idProd = $produto->getId();
+var_dump("Create: $idProd");
+
+
 
 // Read
 $produto = $entityManager->find('Product', $idProd);
-var_dump($produto->getName());
+
+if ($produto === null) {
+    echo "No product found.\n";
+    exit(1);
+}
+
+var_dump("Read: " . $produto->getName());
+
+
 
 // Update
 $produto = $entityManager->find('Product', $idProd);
+
+if ($produto === null) {
+    echo "Product $idProd does not exist.\n";
+    exit(1);
+}
+
 $produto->setName("Produto editado");
 $entityManager->flush();
+var_dump("Update: " . $produto->getName());
+
+
 
 // Delete
 $produto = $entityManager->find('Product', $idProd);
+
+if ($produto === null) {
+    echo "Product $idProd does not exist.\n";
+    exit(1);
+}
+
 $entityManager->remove($produto);
 $entityManager->flush();
+var_dump("Delete");
